requirements.txt
torch>=2.2.0
torchvision>=0.17.0
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
seaborn>=0.13.0
tqdm>=4.66.0
nltk>=3.8.1
rouge-score>=0.1.2
networkx>=3.2.1
pyyaml>=6.0.1
Pillow>=10.0.0
